% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_odds.R
\name{log_odds}
\alias{log_odds}
\title{Calculate the log odds ratio with wald confidence intervals}
\usage{
log_odds(n00, n01, n10, n11, alpha = 0.05, print.result = "df")
}
\arguments{
\item{n00}{A scalar for the number of observations exposed with the condition.}

\item{n01}{A scalar for the number of observations exposed without the
condition.}

\item{n10}{A scalar for the number of observations not exposed with the
condition.}

\item{n11}{A scalar for the number of observations not exposed without the
condition.}

\item{alpha}{Numeric value defining the probability for the wald confidence
interval. Default is an alpha of 0.05, corresponding to a 0.95 \% confidence
interval.}

\item{print.result}{Character. Either "vector" for output given in a vector or
"df" for output given within a tibble (default).}
}
\description{
\code{log_odds} computes the log odds ratio between two binary outcomes.
}
\examples{
# If we have a hypothetical group of smokers (exposed) and non-smokers (not exposed),
# then we can look for the rate of lung cancer (condition).
# 20 smokers have lung cancer (n00),
# 80 smokers do not have lung cancer (n01),
# one non-smoker has lung cancer (n10),
# and 99 non-smokers do not have lung cancer (n11).
log_odds(20, 80, 1, 99)
}
