% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_change.R
\name{perc_change}
\alias{perc_change}
\title{Calculate percentage change}
\usage{
perc_change(old.value, new.value, print.result = "value")
}
\arguments{
\item{old.value}{(numeric) The baseline for the comparison.}

\item{new.value}{(numeric) The value which is compared to the baseline.}

\item{print.result}{(character) Either "value", "percentage", or "text".}
}
\description{
\code{perc_change} returns the percentage change between two values, where one
value (\code{old.value}) is used as a baseline to compare the change with.
}
\details{
The output can be specified via the \code{print.result} parameter. Options are
the raw value, transformed to percentage, or a text description. Default
settings return the raw value.
}
\section{Warning}{
 \code{perc_change} is vectorized and shorter values are
recycled. If the length of \code{old.value} differs from the length of
\code{print.result}, please confirm that the code is doing the desired task
by double-checking with print.result = "text".
}

\examples{
perc_change(7, 5, print.result = "text")
}
