% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term.R
\name{short_term}
\alias{short_term}
\title{Calculate short-term changes for binned data.}
\usage{
short_term(data, value, bin, bin.one = "oldest", mult.observations = FALSE)
}
\arguments{
\item{data}{(data.frame) A data frame or tibble containing the value column
and the bin column}

\item{value}{(unquoted expression) Variable name of the value for which the
short-term changes should be calculated.}

\item{bin}{(unquoted expression) Variable name of the discrete bin numbers.}

\item{bin.one}{(character) Either "oldest" when the first bin (normally 1) is
the oldest out of all bins or "youngest" when the first bin is the most
recent one. "oldest" is the default and corresponds to the convention that
stage 1 is the oldest geologic stage.}

\item{mult.observations}{(logical) Are there more observations per bin?
Defaults to one observation per bin.}
}
\value{
The output is a tibble data frame containing the value column, the
bin column and the calculated short-term change.
}
\description{
Calculate short-term changes for binned data.
}
\section{Details}{

To calculate the short-term change for each row, the coefficient of a linear
regression between the value and bin of interest and each corresponding lag
(the value and bin before the focal bin) is used. This corresponds to a unit
change of the value compared to the previous bin. For example, if we have
20°C at bin 10, and 22°C at bin 11, the short-term temperature change from
bin 10 to 11 is 2. If bin 11 is older than bin 10, one needs to specify this
by setting bin.one = "youngest" and the corresponding short-term temperature
change would hence be calculated from bin 11 to bin 10. If there is more than one observation per bin, and
mult.observations is set to TRUE, the regression is based on all values per
bin.
}

\examples{
dfr <- data.frame(x = rnorm(10), stg = 1:10)
short_term(data = dfr, value = x, bin = stg)
}
